changes updated

+++ app/Http/Controllers/FavouritesongController.php
@@ -8,40 +8,44 @@ use Illuminate\Http\Request;
 
 class FavouritesongController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('CheckExpiredToken');
+    }
     public function getsong(Request $request){
         return response()->json([
             'error' => false,
             'message' => 'Success',
-            'songs' => Favouritesong::where('user_id',$request->user_id)->get(),
+            'song' => Favouritesong::where('user_id',auth()->user()->id)
+            ->with('song')
+            ->get(),
         ]);
     }
-    public function storesong(Request $request){
+    public function song(Request $request){
         try{
-            $Favouritesong= new Favouritesong();
-            $Favouritesong->user_id= $request->user_id;
-            $Favouritesong->song_id= $request->song_id;
-            $Favouritesong->save();
-            return response()->json([
-                'error' => false,
-                'message' => 'Success'
-            ]);
-        }
-        catch(Exception $e){
-            return response()->json([
-                'error' => true,
-                'message' => $e->getMessage(),
-            ]);
-        }
-    }
 
-    public function unfavsong(Request $request){
-        try{
-            $Favouritesong= Favouritesong::where('song_id',$request->song_id)->first();
-            $Favouritesong->delete();
-            return response()->json([
-                'error' => false,
-                'message' => 'Success'
-            ]);
+                if($request->favourite == 'true'){
+
+                $Favouritesong= new Favouritesong();
+                $Favouritesong->user_id= auth()->user()->id;
+                $Favouritesong->song_id= $request->song_id;
+                $Favouritesong->save();
+                return response()->json([
+                    'error' => false,
+                    'message' => 'Success'
+                ]);
+
+                }
+                else{
+
+                    $Favouritesong= Favouritesong::where('user_id',auth()->user()->id)->where('song_id',$request->song_id);
+                    $Favouritesong->delete();
+                    return response()->json([
+                        'error' => false,
+                        'message' => 'do Success'
+                    ]);
+                }
+
         }
         catch(Exception $e){
             return response()->json([
